Clean up obsolete partition directories
When changes to a device are committed using parted, we currently do not
remove any obsolete partition directories in /var/lib/partman/devices/$dev.
This stale data is confusing and could potentially lead to problems.

Introduce a new function device_cleanup_partitions() in the commit.sh
function library to ensure the partition info is consistent.

r50654