Remove valid_visuals.d and use visual.d instead
Previously, valid_visuals.d dynamically built a list of scripts in
visual.d that needed to be called to generate a line in choose_partition.

This mechanism only served in two cases:
 - type: primary/extended partions (e.g. msdos table)
 - name: named partitions (e.g. mac table)

It was not extensible for more visual scripts than the ones shipped in
partman-base, as the result was filtered through an hard-coded list
defined in the update.d/visual script.

As scripts were executed or not, depending on the table type, the dynamic
number of columns were varying from a disk to another.  In the context of
multiple table type, the rendered output would really be difficult to
adapt to the new column alignment features of cdebconf.

So, in order to simplify all this, visual.d scripts are now called
directly by update.d/visual, in their order.  The previous checks have
been inlined in the corresponding visual.d script at a little bit of extra
CPU cost.

The unused visual.d/parted_fs was removed along the way.

lib/disk-label.sh in partman-partitioning now longer needs to remove the
"visuals" file and has been updated accordingly.

r54700